Same code refactor, as SourceAdapter, but for Articles adapter
Moving isolated classes into a new packages, holders
